I add item to the table - passed
