+++ PlaywrigthSpecFlow/Features/WebTable/WebTableTests.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace SpecFlow_Playwrigth.Features.WebTable
+namespace PlaywrigthSpecFlow.Features.WebTable
 {
     using TechTalk.SpecFlow;
     using System;
